mitigate over-eager HTML-build & potential for merge-conflict (#2)
== Summary

There was an attempt made to always preview production-ready HTML locally -- rather than development HTML -- and to leave management of site HTML changesets up to the author. However, this proved confusing, frustrating and ripe for merge conflicts, due to the eagerness with which Jekyll writes/updates _all_ HTML artifacts.

Instead, the development environment is simplified, in that the separate Jekyll build server is removed, (and the Jekyll preview server's `_site` directory is again ignored). However, the manage-promote command is made a little more complex, as it must now build production-ready HTML -- now under conspicuous root-level directory `site` -- and ensure that this is committed to the repository, (prior to the subtree-push to the `gh-pages` branch).

== Changes

* ignore and remove _site from repository

builds write unstable metadata (includes datetimes) -- will try ignoring
by default and then force-committing upon build.

* force-add _site build

* re-remove textbook/_site/ from repository

* playing with textbook content

* "textbook content promotion (automatic commit)"

* textbook content promotion (automatic commit)

* abandon promotion of preview HTML and build+commit during promote instead

also enables manage-build command to recover from error of jupyter
server container not running, and in this case run a new temporary
container.