processed draft of textbook chapters 6 and 9
Technical notes
---------------

**Separate data files**

Separate files pose various problems:

* Mainly: They won't be included with the notebook when it's downloaded
  from the textbook site or otherwise interacted with by the
  reader/student/user.
* It's sometimes non-obvious to link to them / access them.
* They're easy to forget (to commit).

Unfortunately, Jupyter (Books) doesn't yet have a great solution for
this. (There's a Jupyter Notebook "attachment" feature, but just for
images; and, this doesn't currently work with Jupyter Books.)

In this case, the data isn't enormous, and I thought it reasonable to
include it in the notebook as a string; and, I called this out with a
toggling cell. Certainly there are other possibilities.

(And, data to which we want to refer across notebooks is yet another
matter; though, there are apparently ways to refer to variables across
Notebooks….)

**Table of contents**

The textbook won't build files unless they are included in the table of
contents file: `_toc.yml`.

**Scrolling cell outputs**

Some of the cell outputs were ENORMOUS (many columns or rows). These
were made to scroll in the HTML with the Jupyter cell tag
`output_scroll`.

Editing notes
-------------

I polished the section as I thought reasonable for clarity and
legibility. (But certainly some of this is subjective.)

To note, I ensured that there were proper headers -- (the title is of
course an h1 with `#`, and below that are h2 headers with `##`).